src/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-admin-interface==0.16.3
django-autoslug==1.9.8
django-colorfield==0.4.1
django-countries==7.2.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-phonenumber-field==5.1.0
idna==2.10
memcache==0.5.1
oauthlib==3.1.0
phonenumbers==8.12.23
Pillow==8.2.0
psutil==5.8.0
pycparser==2.20
PyJWT==2.1.0
pymemcache==3.4.3
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.4
